Restructure example component's test-run service and Jupiter test file locations
- Move ExampleTestRunServices.groovy from src/main/groovy to src/test/groovy,
  since it only exists to trigger/poll test runs
- Move the Test Run Services block in services.xml to the end of the file
  and update its location attributes to match
- Move ExampleJupiterTests.groovy up one directory (out of the jupiter/
  subfolder) since the class name already conveys its purpose; update its
  package declaration and the jupiter-test-suite class-name in tests.xml
  to match
